+++ .idea/workspace.xml
@@ -6,15 +6,8 @@
   <component name="ChangeListManager">
     <list default="true" id="8461b64b-0823-49c5-a2c2-b16655903f24" name="Changes" comment="">
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/Old Stuff java/authentication/AuthenticationService.java" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/Old Stuff java/authentication/IAuthenticationService.java" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/Old Stuff java/authentication/User.java" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/Old Stuff java/calendar/CalendarItem.java" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/Old Stuff java/calendar/CalendarList.java" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/Old Stuff java/com/example/cab302javaproject/Authentication.java" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/Old Stuff java/com/example/cab302javaproject/LifeStyleCalendar.java" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/Old Stuff java/com/example/cab302javaproject/LifeStyleCalendarController.java" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/Old Stuff java/module-info.java" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/com/example/cab302javaproject/LifeStyleCalendar.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/com/example/cab302javaproject/LifeStyleCalendar.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/com/example/cab302javaproject/UserDetails.java" beforeDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -62,18 +55,18 @@
     <option name="hideEmptyMiddlePackages" value="true" />
     <option name="showLibraryContents" value="true" />
   </component>
-  <component name="PropertiesComponent"><![CDATA[{
-  "keyToString": {
-    "ASKED_ADD_EXTERNAL_FILES": "true",
-    "Application.LifeStyleCalendar.executor": "Run",
-    "JUnit.All in CAB302JavaProject.executor": "Run",
-    "RunOnceActivity.OpenProjectViewOnStart": "true",
-    "RunOnceActivity.ShowReadmeOnStart": "true",
-    "git-widget-placeholder": "Ryan",
-    "kotlin-language-version-configured": "true",
-    "settings.editor.selected.configurable": "preferences.pluginManager"
+  <component name="PropertiesComponent">{
+  &quot;keyToString&quot;: {
+    &quot;ASKED_ADD_EXTERNAL_FILES&quot;: &quot;true&quot;,
+    &quot;Application.LifeStyleCalendar.executor&quot;: &quot;Run&quot;,
+    &quot;JUnit.All in CAB302JavaProject.executor&quot;: &quot;Run&quot;,
+    &quot;RunOnceActivity.OpenProjectViewOnStart&quot;: &quot;true&quot;,
+    &quot;RunOnceActivity.ShowReadmeOnStart&quot;: &quot;true&quot;,
+    &quot;git-widget-placeholder&quot;: &quot;Ryan&quot;,
+    &quot;kotlin-language-version-configured&quot;: &quot;true&quot;,
+    &quot;settings.editor.selected.configurable&quot;: &quot;preferences.pluginManager&quot;
   }
-}]]></component>
+}</component>
   <component name="RunManager" selected="Application.LifeStyleCalendar">
     <configuration name="LifeStyleCalendar" type="Application" factoryName="Application" temporary="true" nameIsGenerated="true">
       <option name="MAIN_CLASS_NAME" value="com.example.cab302javaproject.LifeStyleCalendar" />
