--- src/main/java/com/example/cab302javaproject/UserDetails.java
@@ -1,50 +1,0 @@
-package com.example.cab302javaproject;
-
-import java.util.UUID;
-
-public class UserDetails {
-    private UUID uuid;
-    private String name;
-    private String email;
-    private String password;
-
-    public UserDetails(UUID uuid, String name, String email, String password) {
-        this.uuid = uuid;
-        this.name = name;
-        this.email = email;
-        this.password = password;
-    }
-
-    // Getters and Setters
-    public UUID getUuid() {
-        return uuid;
-    }
-
-    public void setUuid(UUID uuid) {
-        this.uuid = uuid;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-}
